feat: Add MySQL and PostgreSQL support to DataAccess
- Updated PagedResult to use long for TotalCount.
- Added MySQL and PostgreSQL projects to the solution.
- Implemented ClassMapRegistration for MySQL and PostgreSQL to handle table creation.
- Created Registration classes for MySQL and PostgreSQL to configure services.
- Developed DataAccessContext for MySQL and PostgreSQL to manage database operations.
- Updated README to include new MySQL and PostgreSQL modules.

+++ DataAccess.Abstractions/Models/PagedResult.cs
@@ -7,7 +7,7 @@ public sealed record PagedResult<T>
 
     public required int PageSize { get; init; }
 
-    public required int TotalCount { get; init; }
+    public required long TotalCount { get; init; }
 
     public int TotalPages => (int)Math.Ceiling((double)TotalCount / PageSize);
